Do not resubmit tasks when executors are lost
The DAGScheduler assumes that shuffle data is lost if an executor is
lost and the Spark external shuffle service is not enabled. However,
external shuffle managers other than the default external shuffle
service can store shuffle data outside of the Spark cluster itself. In
this case, completed map taks are not rerun even if the corresponding
executors are lost.

The new `spark.shuffle.external.enabled` property allows this new
external shuffle behavior to be toggled (independently off the built-in
Spark external shuffle service).

Change-Id: Icad0ed1ddee77e39094765dc40814a90b822bc7c